use data file as container for LED patterns

+++ app/include/devices/WS2812.h
@@ -27,13 +27,13 @@ public:
   }
   bool setPattern(const uint32_t* data, size_t count);
 
-  uint getLEDCount() const { return _num_leds; }
+  size_t getLEDCount() const { return _num_leds; }
 
 private:
   std::shared_ptr<PIODevice> _pio;
   uint8_t _pin;
   uint8_t _bits_per_pixel;
-  uint _num_leds;
+  size_t _num_leds;
   std::string _name;
 
   static constexpr int DMA_THRESHOLD = 16;
